requirements.txt
google-genai
python-dotenv
pymongo
gradio